refactor(metadata): unexport buildFKCascadeQuery, relocate metadata tests home
Deferred cleanup from the #432/#434 decouple review. The review's one finding
was that metadata.BuildFKCascadeQuery had been exported SOLELY so a cmd-layer
integration test could reach it — and that the consistent end-state of "tests
follow their code" is to move that test home and re-privatize the builder.

- Unexport metadata.BuildFKCascadeQuery → buildFKCascadeQuery (rename only:
  def + doc + 2 doc cross-refs + the 1 call in ValidateNoFKCascades; zero
  behavior change).
- Relocate the metadata integration tests from cmd/bintrail/cmd_integration_test.go
  (package main) → internal/metadata/metadata_integration_test.go (package
  metadata), beside the existing TakeSnapshot/NewResolver integration tests:
  TestValidateBinlogFormat_row, TestValidateBinlogRowImage_full, the 7
  TestValidateNoFKCascades_* + the unscopedFKCascadeSchemas helper (the reason
  the builder needed reaching — it asserts the raw query's flagged-schema set,
  which ValidateNoFKCascades's pass/fail return can't express; now calls the
  private buildFKCascadeQuery in-package), and the 3 TestEnsureResolver_*.
  Verbatim move: only the metadata. qualifier dropped, the rename, and section
  banners capitalized. The 2 TestBuildFKCascadeQuery_* UNIT tests already lived
  in metadata_test.go (names keep capital B; calls use the lowercase private name).
- Drop the now-orphan internal/metadata import from cmd_integration_test.go.
- Comment scrubs: streamrun_integration_test.go "The cmd-layer ensureResolver
  helper" → "The metadata.EnsureResolver helper".

Net: metadata's public surface shrinks back to exactly production usage; the
query builder is private again; cmd/bintrail no longer carries metadata tests.

Verified: go build/vet clean (unit + integration); full unit suite green
(incl. TestBuildFKCascadeQuery_* against the now-private builder); all 12
relocated integration tests pass in their new home (internal/metadata) against
MySQL; cmd/bintrail integration package still compiles + passes; git grep finds
zero references to the exported symbol (only the two test-function names).
